zmieniłem MarketProductShoppinSession na Listing, bo tamta nazwa jest za długa i rails sypał errorami. Dodałem do Listing kolumnę amount stwierdzającą ilość sztuk produktu i dokończyłem listy zakupowe. Mam skrypt .rb działający na sucho który scrapuje, na dniach mam nadzieję go zaimplementować ale muszę nadrobić pisanie klas do tych serwisów :)

+++ db/migrate/20151027101942_add_barcode_to_products.rb
@@ -1,5 +1,5 @@
 class AddBarcodeToProducts < ActiveRecord::Migration
   def change
-    add_column :products, :barcode, :integer
+    add_column :products, :barcode, :float
   end
 end
